+++ srunner/scenarioconfigs/scenario_configuration.py
@@ -18,7 +18,8 @@ class ActorConfigurationData(object):
     This is a configuration base class to hold model and transform attributes
     """
 
-    def __init__(self, model, transform, rolename='other', autopilot=False, random=False, amount=1, color=None):
+    def __init__(self, model, transform, rolename='other', autopilot=False,
+                 random=False, amount=1, color=None, category="car"):
         self.model = model
         self.rolename = rolename
         self.transform = transform
@@ -26,6 +27,7 @@ class ActorConfigurationData(object):
         self.random_location = random
         self.amount = amount
         self.color = color
+        self.category = category
 
 
 class ActorConfiguration(ActorConfigurationData):
@@ -38,7 +40,7 @@ class ActorConfiguration(ActorConfigurationData):
     """
 
     def __init__(self, node, rolename):
-        
+
         pos_x = float(node.attrib.get('x', 0))
         pos_y = float(node.attrib.get('y', 0))
         pos_z = float(node.attrib.get('z', 0))